refactor(archiving): use archiving utils from `archives.py`
